refactor: Move `ShinyHTMLTextDocument` to `shiny.ui`
Exporting it at the top level of `shiny` alongside `App` overstated its
importance -- it is a UI component, and the page containers it stands in for
(`ui.page_fluid()` and friends) all live in `shiny.ui`.

Also add it to `_known_missing_express_ui`: Express builds the page itself, so
an author-owned document has no Express counterpart.
