+++ shiny/__init__.py
@@ -17,7 +17,7 @@ from . import session
 from . import ui
 
 # Private submodules that have some user-facing functionality
-from ._app import App, ShinyHTMLTextDocument
+from ._app import App
 from ._validation import req
 from ._deprecated import *
 
@@ -50,7 +50,6 @@ __all__ = (
     "otel",
     # _app.py
     "App",
-    "ShinyHTMLTextDocument",
     # _main.py
     "run_app",
     # _modules.py
