+++ shiny/express/ui/__init__.py
@@ -359,6 +359,8 @@ _known_missing_express_ui = {
         "page_fluid",
         "page_auto",
         "page_output",
+        # The whole page is an author-owned document, so Express has nothing to build
+        "ShinyHTMLTextDocument",
         "showcase_bottom",
         "showcase_left_center",
         "showcase_top_right",
