+++ shiny/ui/__init__.py
@@ -69,6 +69,7 @@ from ._card import (
 )
 from ._chat import Chat, chat_ui
 from ._download_button import download_button, download_link
+from ._html_text_document import ShinyHTMLTextDocument
 from ._include_helpers import include_css, include_js
 from ._input_action_button import input_action_button, input_action_link
 from ._input_check_radio import (
@@ -400,6 +401,7 @@ __all__ = (
     "TagAttrValue",
     "tags",
     "HTML",
+    "ShinyHTMLTextDocument",
     "head_content",
     "p",
     "h1",
